requirements.txt
langchain
openai
faiss-cpu
numpy
Pillow
streamlit
python-dotenv
langchain-community